fix datatypes

+++ src/processor.cpp
@@ -4,21 +4,21 @@
 
 // DONE: Return the aggregate CPU utilization
 float Processor::Utilization() {
-  static int previowait, previdle, prevuser, prevnice, prevsystem, previrq,
+  static long previowait, previdle, prevuser, prevnice, prevsystem, previrq,
       prevsoftirq, prevsteal;
-  int iowait, idle, user, nice, system, irq, softirq, steal;
+  long iowait, idle, user, nice, system, irq, softirq, steal;
   float PrevIdle, Idle, PrevNonIdle, NonIdle, PrevTotal, Total, totald, idled;
   static bool IsInit = false;
   float procUtil = 0;
   std::vector<std::string> cpuUtil = LinuxParser::CpuUtilization();
-  user = std::stoi(cpuUtil[0]);
-  nice = std::stoi(cpuUtil[1]);
-  system = std::stoi(cpuUtil[2]);
-  idle = std::stoi(cpuUtil[3]);
-  iowait = std::stoi(cpuUtil[4]);
-  irq = std::stoi(cpuUtil[5]);
-  softirq = std::stoi(cpuUtil[6]);
-  steal = std::stoi(cpuUtil[7]);
+  user = std::stol(cpuUtil[0]);
+  nice = std::stol(cpuUtil[1]);
+  system = std::stol(cpuUtil[2]);
+  idle = std::stol(cpuUtil[3]);
+  iowait = std::stol(cpuUtil[4]);
+  irq = std::stol(cpuUtil[5]);
+  softirq = std::stol(cpuUtil[6]);
+  steal = std::stol(cpuUtil[7]);
   // guest = std::stoi(cpuUtil[8]);
   // guest_nice = std::stoi(cpuUtil[9]);
   if (IsInit) {
